[sled-agent] Decouple RSS from Bootstrap Agent, run in distinct task
